+++ src/translators/index.ts
@@ -1,9 +1,9 @@
 export * from './BaseTranslator';
 export * from './Translator';
+export * from './FallbackTranslator';
 export * from './FakeTranslator';
 
 // Stable translators
-export * from './MicrosoftTranslator';
 export * from './GoogleTranslator';
 export * from './DeepLTranslator';
 export * from './TartuNLPTranslator';