+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+# 4.0.0
+
+**Breaking change** - `MicrosoftTranslator` has been removed. Use `GoogleTranslator` instead.
+
+- [#146](https://github.com/translate-tools/core/pull/146) - translator MicrosoftTranslator is deleted
+- [#146](https://github.com/translate-tools/core/pull/146) - added primitive to implement fallbacks to backup translators
+
 # 3.2.2
 
 - [#131](https://github.com/translate-tools/core/pull/131) - fix MicrosoftTranslator errors while token fetching after network problems